feat: add permissions settings components and utilities
