+++ src/components/Tenants/AppSettings/PermissionsSettings/types.ts
@@ -1,0 +1,25 @@
+export type ToggleAfterChangeHandler = (
+    fieldPath: string | string[],
+    value: boolean,
+    currentFormData?: { settings?: Record<string, unknown> },
+) => void;
+
+export type ChatTypeCardKey = 'oneOnOne' | 'liveChat' | 'group' | 'groupInternal';
+
+export type PermissionsSettingsCommonArgs = {
+    tenantId: string;
+    /** Hide chat-type cards by key (e.g. liveChat on super-admin settings — managed under Global Configs). */
+    excludeCardKeys?: Array<ChatTypeCardKey>;
+};
+
+export type ChatTypeCardDef = {
+    key: ChatTypeCardKey;
+    titleKey: string;
+    descriptionKey: string;
+    Icon: import('react').FunctionComponent<import('react').SVGProps<SVGSVGElement>>;
+    masterField?: string[];
+    toggles: Array<{
+        labelKey: string;
+        field: string[];
+    }>;
+};
